cherryLeafClassification
Another image classification project which classifies the situations of cherry leaf is added. In this project, dataset is used from Kaggle.com and CNN architecture used to train model. Adam algorithm was very important to define the weights of the neurons. Three callbacks used to prevent overfitting which are EarlyStopping, ReduceLROnPlateau and ModelCheckpoint.